+++ spec/browser-runner.js
@@ -1,11 +1,8 @@
 $(function() {
   var jasmineEnv = jasmine.getEnv();
-  jasmineEnv.updateInterval = 250;
-
-  jasmine.getFixtures().fixturesPath =
-  jasmine.getStyleFixtures().fixturesPath = 'spec/fixtures';
-
-  var htmlReporter = new jasmine.HtmlReporter();
+  var jsReporter = new jasmine.JSReporter(),
+      htmlReporter = new jasmine.HtmlReporter();
+  jasmineEnv.addReporter(jsReporter);
   jasmineEnv.addReporter(htmlReporter);
 
   jasmineEnv.specFilter = function(spec) {