+++ spec/callbacksSpec.js
@@ -2,7 +2,6 @@ describe("Dragdealer callbacks", function() {
 
   beforeEach(function() {
     this.addMatchers(matchers);
-    jasmine.Clock.useMock();
   });
 
   describe("should call regular callback", function() {
@@ -95,7 +94,7 @@ describe("Dragdealer callbacks", function() {
       expect(callback).toHaveBeenCalledWith(0.075, 0.05);
 
       helpers.drop('square-slider');
-      jasmine.Clock.tick(3000);
+      helpers.callRequestAnimationFrameMock(3000);
       // We don't care about the number of calls, just the last call
       expect(callback).toHaveBeenCalledWith(0.375, 0.25);
     });
@@ -119,7 +118,7 @@ describe("Dragdealer callbacks", function() {
       expect(callback).toHaveBeenCalledWith(0.1, 0);
 
       helpers.drop('simple-slider');
-      jasmine.Clock.tick(3000);
+      helpers.callRequestAnimationFrameMock(3000);
       // We don't care about the number of calls, just the last call
       expect(callback).toHaveBeenCalledWith(0.5, 0);
     });
@@ -138,9 +137,74 @@ describe("Dragdealer callbacks", function() {
     expect(callback).toHaveBeenCalledWith(0.04, 0);
 
     helpers.drop('masked-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     // We want to make sure the value has reached 0.2 and is not something like
     // 0.19999948332063716
     expect(callback).toHaveBeenCalledWith(0.2, 0);
   });
+
+  describe("should call dragstart callback", function() {
+
+    it("with inital values at drag start", function() {
+      var dragStartCallback = jasmine.createSpy(),
+          dragdealer = helpers.initDragdealer('square-slider', {
+        dragStartCallback: dragStartCallback
+      });
+      helpers.dragTo('square-slider', 200, 100);
+
+      expect(dragStartCallback.calls.length).toEqual(1);
+      expect(dragStartCallback).toHaveBeenCalledWith(0, 0);
+    });
+
+  });
+
+  describe("should call dragstop callback", function() {
+
+    it("with exact values on drop without slide", function() {
+      var dragStopCallback = jasmine.createSpy(),
+          dragdealer = helpers.initDragdealer('square-slider', {
+        dragStopCallback: dragStopCallback,
+        slide: false
+      });
+      helpers.dragTo('square-slider', 200, 100);
+      helpers.drop('square-slider');
+
+      expect(dragStopCallback.calls.length).toEqual(1);
+      expect(dragStopCallback).toHaveBeenCalledWith(0.5, 0.25, [0.5, 0.25]);
+    });
+
+    it("with projected values on drop with slide", function() {
+      var dragStopCallback = jasmine.createSpy(),
+          dragdealer = helpers.initDragdealer('square-slider', {
+        dragStopCallback: dragStopCallback,
+        slide: true
+      });
+      // will slide to 100px, 50px
+      helpers.dragTo('square-slider', 20, 10);
+      helpers.drop('square-slider');
+
+      expect(dragStopCallback.calls.length).toEqual(1);
+      expect(dragStopCallback).toHaveBeenCalledWith(0.25, 0.125, [0.05, 0.025]);
+    });
+
+    it("with projected values on drop with slide to steps", function() {
+      var dragStopCallback = jasmine.createSpy(),
+          dragdealer = helpers.initDragdealer('simple-slider', {
+        dragStopCallback: dragStopCallback,
+        // Considering the simple slider has a wrapper of 500px width and a
+        // handle of 100px width, the step positions will be 0, 80, 160, 240,
+        // 320 and 400
+        steps: 6,
+        slide: true
+      });
+      // would slide to 150px, so will slide to 160px step
+      helpers.dragTo('simple-slider', 30, 0);
+      helpers.drop('simple-slider');
+
+      expect(dragStopCallback.calls.length).toEqual(1);
+      expect(dragStopCallback).toHaveBeenCalledWith(0.4, 0, [0.075, 0]);
+    });
+
+  });
+
 });