+++ spec/draggingSpec.js
@@ -2,7 +2,6 @@ describe("Dragging Dragdealer", function() {
 
   beforeEach(function() {
     this.addMatchers(matchers);
-    jasmine.Clock.useMock();
   });
 
   it("should move handle along with mouse after pressing", function() {
@@ -132,7 +131,8 @@ describe("Dragging Dragdealer", function() {
     // of it with every interval loop, normally we'd never move an entire 50px
     // with one mouse move
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
+
     expect('simple-slider').toHavePosition(250, 0);
   });
 
@@ -147,22 +147,22 @@ describe("Dragging Dragdealer", function() {
 
     helpers.dragTo('simple-slider', 100, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(80, 0);
 
     helpers.dragTo('simple-slider', 50, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(80, 0);
 
     helpers.dragTo('simple-slider', 350, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(320, 0);
 
     helpers.dragTo('simple-slider', 210, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(240, 0);
   });
 
@@ -179,25 +179,25 @@ describe("Dragging Dragdealer", function() {
     // is dragged 25px to the right, and will slide 125px, to 125, 0
     helpers.dragTo('simple-slider', 25, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(160, 0);
 
     // is dragged 15px to the left, and will slide 75px, to 135, 0
     helpers.dragTo('simple-slider', 155, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(160, 0);
 
     // is dragged 25px to the right, and will slide 125px, to 285, 0
     helpers.dragTo('simple-slider', 185, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(320, 0);
 
     // is dragged 20px to the left, and will slide 250px, to 70, 0
     helpers.dragTo('simple-slider', 270, 0);
     helpers.drop('simple-slider');
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(80, 0);
   });
 
@@ -239,7 +239,7 @@ describe("Dragging Dragdealer", function() {
     helpers.dragTo('simple-slider', -100, 0);
     helpers.drop('simple-slider');
     expect('simple-slider').toHavePosition(-25, 0);
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(0, 0);
 
     // This goes outside the wrapper with 200px, since 400px is the rightmost
@@ -248,7 +248,7 @@ describe("Dragging Dragdealer", function() {
     helpers.dragTo('simple-slider', 600, 0);
     helpers.drop('simple-slider');
     expect('simple-slider').toHavePosition(450, 0);
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('simple-slider').toHavePosition(400, 0);
   });
 
@@ -263,13 +263,13 @@ describe("Dragging Dragdealer", function() {
     helpers.dragTo('masked-slider', 100, 200);
     helpers.drop('masked-slider');
     expect('masked-slider').toHavePosition(25, 50);
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('masked-slider').toHavePosition(0, 0);
 
     helpers.dragTo('masked-slider', -2000, -1000);
     helpers.drop('masked-slider');
     expect('masked-slider').toHavePosition(-875, -625);
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     expect('masked-slider').toHavePosition(-500, -500);
   });
 
@@ -302,7 +302,7 @@ describe("Dragging Dragdealer", function() {
     helpers.dragTo('simple-slider', 200, 0);
     helpers.drop('simple-slider');
     dragdealer.unbindEventListeners();
-    jasmine.Clock.tick(3000);
+    helpers.callRequestAnimationFrameMock(3000);
     // The handle would reach the 400, 0 position if we wouldn't unbind it
     expect('simple-slider').toHavePosition(200, 0);
   });