+++ spec/helpers.js
@@ -1,9 +1,11 @@
 var helpers = {
 
   initDragdealer: function(dragdealerId, options) {
+    if (!options) options = {};
     loadFixtures(dragdealerId + '.html');
     loadStyleFixtures(dragdealerId + '.css');
-
+    options.customRequestAnimationFrame = this.createRequestAnimationFrameMock();
+    options.customCancelAnimationFrame = this.createCancelAnimationFrameMock();
     return new Dragdealer(dragdealerId, options);
   },
 
@@ -13,21 +15,24 @@ var helpers = {
         wrapperPosition = $wrapper.offset(),
         handlePosition = $handle.offset();
 
+    this.callRequestAnimationFrameMock(0);
+
     // Move to current handle position and press
-    $handle.simulate('mousemove', {
+    $(document).simulate('mousemove', {
       clientX: handlePosition.left,
       clientY: handlePosition.top
     });
 
     $handle.simulate('mousedown');
-    $handle.simulate('mousemove', {
+
+    $(document).simulate('mousemove', {
       clientX: wrapperPosition.left + x,
       clientY: wrapperPosition.top + y
     });
 
     // Dragdealer internal animation delay is 25ms (this should be fixed and
     // dragdealer should be updated instantly on mousemove or mouseup)
-    jasmine.Clock.tick(25);
+    this.callRequestAnimationFrameMock(25);
   },
 
   drop: function(dragdealerId, x, y, handleClass) {
@@ -35,25 +40,28 @@ var helpers = {
     $handle.simulate('mouseup');
   },
 
-  touchDragTo: function (dragdealerId, x, y, handleClass) {
+  touchDragTo: function(dragdealerId, x, y, handleClass) {
     var $wrapper = $('#' + dragdealerId),
         $handle = $wrapper.find('.' + (handleClass || 'handle')),
         wrapperPosition = $wrapper.offset(),
         handlePosition = $handle.offset(),
         result;
 
+    this.callRequestAnimationFrameMock(0);
+
     // Move to current handle position and start touch
     simulateTouchEvent($handle.get(0), 'touchstart', {
-      clientX: handlePosition.left,
-      clientY: handlePosition.top
-    })
-
-    result = simulateTouchEvent($wrapper.get(0), 'touchmove', {
-      clientX: wrapperPosition.left + x,
-      clientY: wrapperPosition.top + y
+      pageX: handlePosition.left,
+      pageY: handlePosition.top
     });
 
-    jasmine.Clock.tick(25);
+    result = simulateTouchEvent($wrapper.get(0), 'touchmove', {
+      pageX: wrapperPosition.left + x,
+      pageY: wrapperPosition.top + y
+    });
+
+
+    this.callRequestAnimationFrameMock(25);
 
     // Return the result of touchmove event dispatch
     // to check if it was canceled or not
@@ -63,11 +71,37 @@ var helpers = {
   touchDrop: function(dragdealerId, x, y, handleClass) {
     var $handle = $('#' + dragdealerId).find('.' + (handleClass || 'handle'));
     simulateTouchEvent($handle.get(0), 'touchend');
+  },
+
+  createRequestAnimationFrameMock: function() {
+    var self = this;
+    this.time = 0;
+    this.memFunc = function() {};
+    return function mockAnimationFrame(func) {
+      self.memFunc = func;
+    };
+  },
+
+  createCancelAnimationFrameMock: function() {
+    var self = this;
+    return function mockCancelAnimationFrame(func) {
+      self.memFunc = function() {};
+    };
+  },
+
+  callRequestAnimationFrameMock: function(milliseconds) {
+    if (this.time === 0) {
+      this.memFunc(0);
+    }
+    for (var t = 25; t <= milliseconds; t += 25) {
+      this.memFunc(this.time + t);
+    }
+    this.time += t;
   }
 
 };
 
-function simulateTouchEvent (element, type, touchOptions) {
+function simulateTouchEvent(element, type, touchOptions) {
   var event = document.createEvent('UIEvent');
   event.initUIEvent(type, true, type !== 'touchcancel', window, 0);
   if (touchOptions) {
@@ -75,4 +109,3 @@ function simulateTouchEvent (element, type, touchOptions) {
   }
   return element.dispatchEvent(event);
 }
-